requirements.txt
# AI / ML
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
onnx==1.19.1
transformers==4.57.1
accelerate==1.9.0
bitsandbytes==0.46.1
sentence-transformers==5.1.2
optimum==2.0.0

# LangChain
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-core==0.3.76
langchain-deepseek==0.1.2
langchain-huggingface==0.3.1
langchain-ollama==1.0.0
langchain-openai==0.3.33
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11

# Web / API
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
requests==2.32.5
python-multipart==0.0.20

# Data / Math / Utils
numpy==1.26.2
pandas==2.2.3
scipy==1.15.2
scikit-learn==1.7.1
matplotlib==3.10.1
seaborn==0.13.2
tqdm==4.67.1
rich==14.1.0
orjson==3.11.3

# Image / PDF processing
pillow==11.1.0
opencv-python==4.12.0.88
pdf2image==1.17.0
PyMuPDF==1.26.5
python-dotenv==1.1.1

# OCR / Vision
easyocr==1.7.2
paddleocr==3.3.2

# streamlit / Vision
streamlit